+++ CHANGELOG.md
@@ -10,6 +10,20 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 ## Unreleased
 
+## 0.2.0
+
+### Added
+
+- Added entity with offset targeting through `commands.ogle_target_entity_with_offset(entity:Entity, offset: Vec2)`.
+- Added `commands.ogle_target(target: OgleTarget)`.
+- Added `commands.ogle_freeze()`.
+- Added `commands.ogle_follow()`.
+- Added `commands.ogle_pancam()`.
+
+### Changed
+
+- Changed `commands.ogle_change_mode` to `commands.ogle_mode`.
+
 ## 0.1.0
 
 - Initial release